adding the next information done this week
its all a lot of iffs. but most of it will be how it is visualised.

later characters can have visuals, sam e as planets.

maybe to showcase the ship where everyone is.

step by step we'll take it there. I think now its important wiht the location, to visualise ev erything and have options to choose where the crew is.