create the systems, planets and moons
later we ade is its habitable (which is based on how its placed.

there are tow ways. we can create a affinity to heat to create a planet and race. if success. we create a way to see if they explored the universe or not.

this can be done in droves. like periods in time they're in.

or we cna just replicate that feeling. although, thats not so much fun coding.

it is possible to have every planet be habbitable. but we make sure to make it rare.

not every habitable planet has a race on it.

the map is 10000 by 10000 but maybe it can be bigger